requirements.txt
openai
streamlit
dotenv
google-cloud-translate==2.0.1